IMP: "Admin" classes refactored. IMP: Admin Pages code refactored (improved). CHG: "Democracy_Poll" class renamed to "Plugin" and moved under namespace. CHG: "Options_CSS" class added to work with "democracy_css" option.
