+++ classes/Admin/Admin_Subpage_Interface.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace DemocracyPoll\Admin;
+
+interface Admin_Subpage_Interface {
+
+	public function __construct( Admin_Page $admin_page );
+	public function load();
+	public function request_handler();
+	public function render();
+
+}
